+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Development
+
+### Added
+- Add `Container#removeComponents()` to remove all child components of a container
+
 ## [2.1.1]
 
 ### Fixed